requirements.txt
# Core dependencies
pyyaml>=6.0
pydantic>=2.0.0

# PAL (Provider Abstraction Layer)
openai>=1.0.0
anthropic>=0.18.0
httpx>=0.25.0

# Optional: For future enhancements
# transformers>=4.35.0
# torch>=2.0.0
